move Publisher -> rla persist clean up cases tests
