+++ cases/src/main/kotlin/org/cryptobiotic/rlauxe/boulder/BoulderElectionFromCvrs.kt
@@ -6,6 +6,7 @@ import org.cryptobiotic.rlauxe.dominion.readDominionCvrExport
 import org.cryptobiotic.rlauxe.audit.*
 import org.cryptobiotic.rlauxe.core.*
 import org.cryptobiotic.rlauxe.dominion.CastVoteRecord
+import org.cryptobiotic.rlauxe.persist.Publisher
 import org.cryptobiotic.rlauxe.persist.csv.writeAuditableCardCsvFile
 import org.cryptobiotic.rlauxe.persist.json.*
 import org.cryptobiotic.rlauxe.raire.*
