+++ rla/src/main/kotlin/org/cryptobiotic/rlauxe/persist/MvrManagerFromRecord.kt
@@ -6,7 +6,6 @@ import com.github.michaelbull.result.unwrap
 import org.cryptobiotic.rlauxe.audit.*
 import org.cryptobiotic.rlauxe.core.*
 import org.cryptobiotic.rlauxe.persist.csv.readCardsCsvIterator
-import org.cryptobiotic.rlauxe.persist.json.Publisher
 import org.cryptobiotic.rlauxe.persist.json.readSampleNumbersJsonFile
 import java.nio.file.Files
 import java.nio.file.Path
