+++ rla/src/main/kotlin/org/cryptobiotic/rlauxe/persist/PersistentAudit.kt
@@ -3,7 +3,6 @@ package org.cryptobiotic.rlauxe.persist
 import org.cryptobiotic.rlauxe.audit.*
 import org.cryptobiotic.rlauxe.core.*
 import org.cryptobiotic.rlauxe.persist.csv.writeAuditableCardCsvFile
-import org.cryptobiotic.rlauxe.persist.json.Publisher
 import org.cryptobiotic.rlauxe.persist.json.writeAuditRoundJsonFile
 import org.cryptobiotic.rlauxe.persist.json.writeSampleNumbersJsonFile
 import org.cryptobiotic.rlauxe.workflow.*
