+++ rla/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/TestPersistentWorkflowPolling.kt
@@ -7,7 +7,7 @@ import org.cryptobiotic.rlauxe.estimate.MultiContestTestData
 import org.cryptobiotic.rlauxe.estimate.makeFuzzedCvrsFrom
 import org.cryptobiotic.rlauxe.persist.PersistentAudit
 import org.cryptobiotic.rlauxe.persist.csv.writeAuditableCardCsvFile
-import org.cryptobiotic.rlauxe.persist.json.Publisher
+import org.cryptobiotic.rlauxe.persist.Publisher
 import kotlin.test.Test
 
 class TestPersistentWorkflowPolling {
